Add openrtb plugin that is really a mopub one.
Two body templates for mopub bid request are also added along with
an initial (also poor) README. Some of the configuration of the
openrtb plugin are in the file. The idea is in a future make it
more accesible (maybe in the config file) and not global as they
are now.